requirements.txt
PyGithub==2.1.1
click==8.1.7
tabulate==0.9.0
python-dotenv==1.0.0
rich==13.7.0
PyYAML==6.0.1 